FIX adapter folder structure
