--- adapters/grpc.go
@@ -1,11 +1,0 @@
-package adapters
-
-// func GrpcService(pool *drivers.DriverGrpc) {
-// 	pool.AddService("transaction", "127.0.0.1", "5100")
-// 	pool.AddService("account", "127.0.0.1", "5100")
-// }
-
-// func GrpcClient(pool *drivers.DriverGrpc) {
-// 	ApiAccount.Client = protobuf.NewApiAccountClient(pool.GetService("account"))
-// 	ApiTransaction.Client = protobuf.NewTransactionServiceClient(pool.GetService("transaction"))
-// }
