--- adapters/messages/audit-trail.go
@@ -1,1 +1,0 @@
-package messages
