--- adapters/messages/transaction.go
@@ -1,1 +1,0 @@
-package messages
