--- adapters/rabbit.go
@@ -1,12 +1,0 @@
-package adapters
-
-// import (
-// 	"lets-go-framework/adapters/clients"
-// 	"lets-go-framework/lets/frameworks"
-// )
-
-// func RabbitRegister(r frameworks.RabbitClient) {
-// 	clients.RabbitTransfer.Driver = r
-// 	clients.RabbitBalance.Driver = r
-// 	clients.RabbitNotification.Driver = r
-// }
