+++ app/adapters/clients/RabbitTransfer.go
@@ -2,22 +2,22 @@ package clients
 
 import (
 	"fmt"
-	"lets-go-framework/adapters/data"
+	"lets-go-framework/app/adapters/data"
 	"lets-go-framework/lets/rabbitmq"
 	"os"
 )
 
-var RabbitBalance rabbitBalance
+var RabbitTransfer = rabbitTransfer{}
 
-type rabbitBalance struct {
+type rabbitTransfer struct {
 	Driver rabbitmq.RabbitClient
 }
 
-func (r *rabbitBalance) BalanceTransfer(correlationId string, data *data.EventTransfer) error {
+func (r *rabbitTransfer) Transfer(correlationId string, data *data.EventTransfer) error {
 	rabbit := r.Driver
 
 	event := rabbitmq.Event{
-		Name:          "balance-transfer",
+		Name:          "transfer",
 		Data:          data,
 		CorrelationId: correlationId,
 		Exchange:      rabbit.GetDst().GetExchange(),
@@ -37,11 +37,11 @@ func (r *rabbitBalance) BalanceTransfer(correlationId string, data *data.EventTr
 	return nil
 }
 
-func (r *rabbitBalance) BalanceRollback(data *data.EventTransfer) error {
+func (r *rabbitTransfer) TransferRollback(data *data.EventTransferRollback) error {
 	rabbit := r.Driver
 
 	event := rabbitmq.Event{
-		Name:       "balance-transfer-rollback",
+		Name:       "transfer-rollback",
 		Data:       data,
 		Exchange:   rabbit.GetDst().GetExchange(),
 		RoutingKey: os.Getenv("RQ_ROUTING_KEY_TRANSFER"),
