+++ app/adapters/protobuf/service-transaction/api.transaction.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 import "format.transaction.proto";
 package service_transaction;
-option go_package="adapters/protobuf";
+option go_package="app/adapters/protobuf";
 
 message RequestGetTransaction {
  FilterTransaction filter = 1;
