+++ app/adapters/protobuf/service-transaction/format.transaction.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package service_transaction;
-option go_package="adapters/protobuf";
+option go_package="app/adapters/protobuf";
 
 message Transaction {
  int32 id = 1;
